@@ -12,9 +12,10 @@
     "@heroicons/react": "^2.2.0",
     "@heroui/avatar": "^2.2.20",
     "@heroui/react": "^2.8.2",
+    "@swc/helpers": "^0.5.15",
     "framer-motion": "^12.23.12",
     "lucide-react": "^0.540.0",
-    "next": "15.4.8",
+    "next": "^16.1.1",
     "postcss": "^8.5.6",
     "react": "19.1.0",
     "react-dom": "19.1.0",
@@ -28,7 +29,7 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
-    "eslint-config-next": "15.4.2",
+    "eslint-config-next": "^16.1.1",
     "tailwindcss": "^4.1.11",
     "typescript": "^5"
   }
